+++ device-bootloaders/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [SW-14061] Moving the GET_SP_STATS handling to process_performance_request().
 - [SW-14061] Separating voltages read from PVT and PMIC.
 - [SW-14061] Separated the PMIC and SOC power registers.
+- [SW-14061] Reading ASIC voltage for Maxion from PVT.
 ### Deprecated
 ### Removed
 ### Fixed